[persist_redesign] Move `EntryIter` into it's own file
